news_sentiment_feedback/requirements.txt
asgiref==3.7.2
attrs==23.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
Django==4.2.11
exceptiongroup==1.2.0
filelock==3.13.1
fsspec==2024.2.0
h11==0.14.0
huggingface-hub==0.21.4
idna==3.6
lxml==5.1.0
numpy==1.24.4
outcome==1.3.0.post0
packaging==23.2
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tokenizers==0.15.2
tqdm==4.66.2
transformers==4.38.2
trio==0.24.0
trio-websocket==0.11.1
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
webdriver-manager==4.0.1
wsproto==1.2.0
youtube-dl==2021.12.17